[#188] refactor: 좋아요 validate LikeQueryService로 로직 이동
