+++ src/main/java/com/konggogi/veganlife/post/service/PostQueryService.java
@@ -3,12 +3,8 @@ package com.konggogi.veganlife.post.service;
 
 import com.konggogi.veganlife.global.exception.ErrorCode;
 import com.konggogi.veganlife.global.exception.NotFoundEntityException;
-import com.konggogi.veganlife.like.domain.PostLike;
-import com.konggogi.veganlife.like.exception.IllegalLikeStatusException;
-import com.konggogi.veganlife.like.repository.PostLikeRepository;
 import com.konggogi.veganlife.post.domain.Post;
 import com.konggogi.veganlife.post.repository.PostRepository;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,31 +14,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class PostQueryService {
     private final PostRepository postRepository;
-    private final PostLikeRepository postLikeRepository;
 
     public Post search(Long postId) {
         return postRepository
                 .findById(postId)
                 .orElseThrow(() -> new NotFoundEntityException(ErrorCode.NOT_FOUND_POST));
-    }
-
-    public Optional<PostLike> searchPostLike(Long memberId, Long postId) {
-        return postLikeRepository.findByMemberIdAndId(memberId, postId);
-    }
-
-    public void validatePostLikeIsExist(Long memberId, Long postId) {
-        searchPostLike(memberId, postId)
-                .ifPresent(
-                        postLike -> {
-                            throw new IllegalLikeStatusException(ErrorCode.ALREADY_LIKED);
-                        });
-    }
-
-    public PostLike validatePostLikeIsNotExist(Long memberId, Long postId) {
-        return searchPostLike(memberId, postId)
-                .orElseThrow(
-                        () -> {
-                            throw new IllegalLikeStatusException(ErrorCode.ALREADY_UNLIKED);
-                        });
     }
 }
